+++ src/components/tomogram/main.tsx
@@ -15,21 +15,26 @@ import {
   AccordionPanel,
   Skeleton,
   VStack,
+  Spacer,
+  Icon,
 } from "@chakra-ui/react";
 import { ImageCard } from "../visualisation/image";
 import { InfoGroup } from "../visualisation/infogroup";
 import { PlotContainer } from "../visualisation/plotContainer";
 import { Motion } from "../motion/motion";
-import { useEffect, useState } from "react";
+import { Suspense, useEffect, useState } from "react";
 import { client } from "../../utils/api/client";
 import { TomogramData, BasePoint, BaseProcessingJobProps, DataConfig } from "../../schema/interfaces";
 import { CTF } from "../ctf/ctf";
 import { Scatter } from "../plots/scatter";
-import { APNGViewer } from "../visualisation/apng";
 import { setImage } from "../../utils/api/response";
 import { components } from "../../schema/main";
 import { ProcessingTitle } from "../visualisation/processingTitle";
 import { parseData } from "../../utils/generic";
+import React from "react";
+import { MdOpenInNew } from "react-icons/md";
+
+const APNGViewer = React.lazy(() => import("../visualisation/apng"));
 
 const tomogramConfig: DataConfig = {
   include: [
@@ -103,6 +108,8 @@ const Tomogram = ({ autoProc, procJob, status, active = false }: BaseProcessingJ
                     <ImageCard height='85%' title='Central Slice' src={sliceImage} />
                     <Button w='100%' mt='1%' height='13%' alignSelf='end' size='sm' onClick={onOpen}>
                       View Movie
+                      <Spacer />
+                      <Icon as={MdOpenInNew}></Icon>
                     </Button>
                   </GridItem>
                   <GridItem height='20vh'>
@@ -132,7 +139,9 @@ const Tomogram = ({ autoProc, procJob, status, active = false }: BaseProcessingJ
             <ModalHeader paddingBottom={0}>Movie</ModalHeader>
             <ModalCloseButton />
             <ModalBody h={{ base: "90vh", md: "60vh" }}>
-              <APNGViewer src={`tomograms/${tomogram.tomogramId}/movie`} />
+              <Suspense>
+                <APNGViewer src={`tomograms/${tomogram.tomogramId}/movie`} />
+              </Suspense>
             </ModalBody>
           </ModalContent>
         </Modal>