+++ src/components/tomogram/reprocessing.test.tsx
@@ -2,7 +2,7 @@ import { waitFor, screen, fireEvent } from "@testing-library/react";
 import { rest } from "msw";
 import { server } from "../../mocks/server";
 import { renderWithProviders } from "../../utils/test-utils";
-import { TomogramReprocessing } from "./reprocessing";
+import TomogramReprocessing from "./reprocessing";
 
 describe("Tomogram Reprocessing", () => {
   it("should not close when not successful", async () => {