+++ CHANGELOG.rst
@@ -11,6 +11,10 @@ v0.6.0 (--/02/2023)
 - 3D visualisation of volume files for SPA
 - 3D classification information
 
+**Changed**
+
+- Fixes inconsistencies between X overflow on Chrome and Firefox
+
 ++++++++++
 v0.5.0 (21/02/2023)
 ++++++++++